adjusting classes for login function

+++ noWaste/src/main/java/A3/project/noWaste/service/impl/UserImpl.java
@@ -1,13 +1,14 @@
 package A3.project.noWaste.service.impl;
 
 import A3.project.noWaste.domain.User;
-import A3.project.noWaste.domain.dto.UserDTO;
+import A3.project.noWaste.dto.UserDTO;
+import A3.project.noWaste.exceptions.ObjectNotFoundException;
 import A3.project.noWaste.infra.UserRepository;
 import A3.project.noWaste.service.UserService;
-import A3.project.noWaste.service.exceptions.DataIntegratyViolationException;
-import A3.project.noWaste.service.exceptions.ObjectNotFoundException;
+import A3.project.noWaste.exceptions.DataIntegratyViolationException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
@@ -22,13 +23,9 @@ public class UserImpl implements UserService {
     @Autowired
     private UserRepository repository;
 
+    @Autowired
+    private PasswordEncoder passwordEncoder;
 
-    // find a user
-    @Override
-    public User findById(Integer Id) {
-        Optional<User> obj = repository.findById(Id);
-        return obj.orElseThrow(() -> new ObjectNotFoundException("Objeto não encontrado"));
-    }
 
     // find all Users
     @Override
@@ -36,25 +33,41 @@ public class UserImpl implements UserService {
         return repository.findAll();
     }
 
+
     // create User
     @Override
     public User create(UserDTO obj) {
         findByEmail(obj);
-        User user = mapper.map(obj, User.class);
-        return repository.save(user);
+
+        User newUser = new User();
+        newUser.setPassword(passwordEncoder.encode(obj.getPassword()));
+        newUser.setUsername(obj.getUsername());
+        newUser.setEmail(obj.getEmail());
+        newUser.setId(obj.getId());
+
+        return repository.save(newUser);
     }
 
     // update user
     @Override
     public User update(UserDTO obj) {
         findByEmail(obj);
+
+        User newUser = new User();
+        newUser.setPassword(passwordEncoder.encode(obj.getPassword()));
+        newUser.setUsername(obj.getUsername());
+        newUser.setEmail(obj.getEmail());
+        newUser.setId(obj.getId());
+
         return repository.save(mapper.map(obj, User.class));
     }
 
     // delete User
     @Override
     public void delete(Integer id) {
-        findById(id);
+        if (!repository.findById(id).isPresent()) {
+            throw new ObjectNotFoundException("Usuário não existe");
+        }
         repository.deleteById(id);
     }
